schema fixes

+++ schemas/schemas.py
@@ -4,6 +4,7 @@ Complete schema definitions for all entities
 """
 
 from datetime import datetime
+from decimal import Decimal
 from typing import Optional, List
 from pydantic import BaseModel, EmailStr, Field
 from enum import Enum
@@ -151,8 +152,8 @@ class DoorTypeShortResponse(DoorTypeBase):
 # Door Type Thickness Option Schemas
 class DoorTypeThicknessOptionBase(BaseModel):
     door_type_id: int
-    thickness_value: float = Field(..., description="Thickness with up to 2 decimal places")
-    cost_per_sqft: float = Field(..., description="Price adjustment with up to 2 decimal places")
+    thickness_value: Decimal = Field(..., description="Thickness with up to 2 decimal places")
+    cost_per_sqft: Decimal = Field(..., description="Price adjustment with up to 2 decimal places")
     is_active: bool = True
 
 
@@ -162,8 +163,8 @@ class DoorTypeThicknessOptionCreate(DoorTypeThicknessOptionBase):
 
 class DoorTypeThicknessOptionUpdate(DoorTypeThicknessOptionBase):
     door_type_id: Optional[int] = None
-    thickness_value: Optional[float] = None
-    cost_per_sqft: Optional[float] = None
+    thickness_value: Optional[Decimal] = None
+    cost_per_sqft: Optional[Decimal] = None
     is_active: Optional[bool] = None
 
 
@@ -184,8 +185,8 @@ class AttributeBase(BaseModel):
     description: Optional[str] = None
     double_side: bool = False
     cost_type: CostType = Field(..., description="Type of cost calculation")
-    fixed_cost: Optional[float] = Field(None, description="Fixed cost with up to 2 decimal places")
-    cost_per_unit: Optional[float] = Field(None, description="Cost per unit with up to 2 decimal places")
+    fixed_cost: Optional[Decimal] = Field(None, description="Fixed cost with up to 2 decimal places")
+    cost_per_unit: Optional[Decimal] = Field(None, description="Cost per unit with up to 2 decimal places")
     unit_id: Optional[int] = None
     is_active: bool = True
 
@@ -199,8 +200,8 @@ class AttributeUpdate(AttributeBase):
     description: Optional[str] = None
     double_side: Optional[bool] = None
     cost_type: Optional[CostType] = None
-    fixed_cost: Optional[float] = None
-    cost_per_unit: Optional[float] = None
+    fixed_cost: Optional[Decimal] = None
+    cost_per_unit: Optional[Decimal] = None
     unit_id: Optional[int] = None
     is_active: Optional[bool] = None
 
@@ -232,8 +233,8 @@ class AttributeListResponse(BaseModel):
     description: Optional[str] = None
     double_side: bool
     cost_type: CostType
-    fixed_cost: Optional[float] = None
-    cost_per_unit: Optional[float] = None
+    fixed_cost: Optional[Decimal] = None
+    cost_per_unit: Optional[Decimal] = None
     unit_id: Optional[int] = None
     is_active: bool
 
@@ -246,8 +247,8 @@ class AttributeOptionBase(BaseModel):
     attribute_id: int
     name: str
     description: Optional[str] = None
-    cost: float = Field(..., description="Cost with up to 2 decimal places")
-    cost_per_unit: Optional[float] = Field(None, description="Cost per unit with up to 2 decimal places")
+    cost: Decimal = Field(..., description="Cost with up to 2 decimal places")
+    cost_per_unit: Optional[Decimal] = Field(None, description="Cost per unit with up to 2 decimal places")
     unit_id: Optional[int] = None
     display_order: int = 0
     is_active: bool = True
@@ -260,8 +261,8 @@ class AttributeOptionCreate(AttributeOptionBase):
 class AttributeOptionUpdate(AttributeOptionBase):
     name: Optional[str] = None
     description: Optional[str] = None
-    cost: Optional[float] = None
-    cost_per_unit: Optional[float] = None
+    cost: Optional[Decimal] = None
+    cost_per_unit: Optional[Decimal] = None
     unit_id: Optional[int] = None
     display_order: Optional[int] = None
     is_active: Optional[bool] = None
@@ -410,7 +411,7 @@ class QuotationUpdate(QuotationBase):
 class QuotationResponse(QuotationBase):
     id: int
     quotation_number: str
-    total_amount: Optional[float] = None
+    total_amount: Optional[Decimal] = None
     customer: "CustomerResponse"
     items: List["QuotationItemResponse"] = []
     created_by: Optional[str] = None
@@ -428,10 +429,10 @@ class QuotationItemBase(BaseModel):
     door_type_id: int
     thickness_option_id: int
     quantity: int = 1
-    length: float = Field(..., description="Length with up to 2 decimal places")
-    breadth: float = Field(..., description="Breadth with up to 2 decimal places")
-    tax_percentage: float = Field(default=float('0'), description="Tax percentage applied to unit price with attributes")
-    discount_amount: float = Field(default=float('0'), description="Flat discount amount per unit after tax")
+    length: Decimal = Field(..., description="Length with up to 2 decimal places")
+    breadth: Decimal = Field(..., description="Breadth with up to 2 decimal places")
+    tax_percentage: Decimal = Field(default=Decimal('0'), description="Tax percentage applied to unit price with attributes")
+    discount_amount: Decimal = Field(default=Decimal('0'), description="Flat discount amount per unit after tax")
 
 
 class QuotationItemCreate(QuotationItemBase):
@@ -444,8 +445,8 @@ class QuotationItemUpdate(QuotationItemBase):
     door_type_id: Optional[int] = None
     thickness_option_id: Optional[int] = None
     quantity: Optional[int] = None
-    length: Optional[float] = None
-    breadth: Optional[float] = None
+    length: Optional[Decimal] = None
+    breadth: Optional[Decimal] = None
 
 
 class QuotationItemResponse(QuotationItemBase):
@@ -453,13 +454,13 @@ class QuotationItemResponse(QuotationItemBase):
     door_type: "DoorTypeShortResponse"
     thickness_option: Optional["DoorTypeThicknessOptionResponse"] = None
     attributes: List["QuotationItemAttributeResponse"] = None
-    base_cost_per_unit: Optional[float] = None
-    attribute_cost_per_unit: Optional[float] = None
-    unit_price_before_tax: Optional[float] = None  # Price after discount, before tax
-    unit_price_with_attributes: Optional[float] = None  # Final price (after discount and tax)
-    total_item_cost: Optional[float] = None
-    tax_percentage: Optional[float] = None
-    discount_amount: Optional[float] = None
+    base_cost_per_unit: Optional[Decimal] = None
+    attribute_cost_per_unit: Optional[Decimal] = None
+    unit_price_before_tax: Optional[Decimal] = None  # Price after discount, before tax
+    unit_price_with_attributes: Optional[Decimal] = None  # Final price (after discount and tax)
+    total_item_cost: Optional[Decimal] = None
+    tax_percentage: Optional[Decimal] = None
+    discount_amount: Optional[Decimal] = None
     created_by: Optional[str] = None
     updated_by: Optional[str] = None
     created_at: Optional[datetime] = None
@@ -475,7 +476,7 @@ class QuotationItemAttributeBase(BaseModel):
     attribute_id: int
     selected_option_id: Optional[int] = None
     double_side: bool = False
-    direct_cost: Optional[float] = Field(None, description="Direct cost entered by user for this attribute")
+    direct_cost: Optional[Decimal] = Field(None, description="Direct cost entered by user for this attribute")
 
 
 class QuotationItemAttributeCreate(QuotationItemAttributeBase):
@@ -487,7 +488,7 @@ class QuotationItemAttributeUpdate(QuotationItemAttributeBase):
     quotation_item_id: Optional[int] = None
     attribute_id: Optional[int] = None
     double_side: Optional[bool] = None
-    direct_cost: Optional[float] = None
+    direct_cost: Optional[Decimal] = None
 
 
 class QuotationItemAttributeResponse(QuotationItemAttributeBase):
@@ -495,8 +496,8 @@ class QuotationItemAttributeResponse(QuotationItemAttributeBase):
     attribute: "AttributeShortResponse"
     selected_option: Optional["AttributeOptionResponse"] = None
     unit_values: List["UnitValueResponse"] = None
-    calculated_cost: Optional[float] = None
-    total_attribute_cost: Optional[float] = None
+    calculated_cost: Optional[Decimal] = None
+    total_attribute_cost: Optional[Decimal] = None
     created_by: Optional[str] = None
     updated_by: Optional[str] = None
     created_at: Optional[datetime] = None
@@ -509,8 +510,8 @@ class QuotationItemAttributeResponse(QuotationItemAttributeBase):
 class UnitValueBase(BaseModel):
     quotation_item_attribute_id: int
     unit_id: int
-    value1: Optional[float] = Field(None, description="Value1 with up to 2 decimal places")
-    value2: Optional[float] = Field(None, description="Value2 with up to 2 decimal places")
+    value1: Optional[Decimal] = Field(None, description="Value1 with up to 2 decimal places")
+    value2: Optional[Decimal] = Field(None, description="Value2 with up to 2 decimal places")
 
 
 class UnitValueCreate(UnitValueBase):
@@ -520,8 +521,8 @@ class UnitValueCreate(UnitValueBase):
 class UnitValueUpdate(UnitValueBase):
     quotation_item_attribute_id: Optional[int] = None
     unit_id: Optional[int] = None
-    value1: Optional[float] = None
-    value2: Optional[float] = None
+    value1: Optional[Decimal] = None
+    value2: Optional[Decimal] = None
 
 
 class UnitValueResponse(UnitValueBase):
